Crate bevy_rust_gpu

bevy-rust-gpu

A set of bevy plugins supporting the use of rust-gpu shaders.

Features include hot-reloading, metadata-based entry point validation, and active entry point export.

Can be used in conjunction with rust-gpu-builder and permutate-macro to drive a real-time shader recompilation pipeline.

Modules

bevy_pbr_rust

bevy-pbr-rust-backed RustGpuMaterial implementation for StandardMaterial.

entry_point_export

Adds support for exporting the active entry point set to a .json file.

prelude

Structs

RustGpu

Extends a Material with rust-gpu shader support.

RustGpuBuilderOutput

Compile output from rust-gpu-builder, includes SPIR-V binary modules and entry point metadata.

RustGpuKey

RustGpu pipeline key.

RustGpuMaterial2dPlugin

Configures backend [Material2d] support for RustGpu<M>.

RustGpuMaterialPlugin

Configures backend [Material] support for RustGpu<M>.

RustGpuPlugin

Main Rust-GPU plugin.

RustGpuSettings

Type-level RustGpu material settings

Enums

RustGpuBuilderModules

Module binary data container.

Traits

EntryPoint

A rust-gpu entry point for use with RustGpuMaterial.

RustGpuMaterial

A [Material] type with statically-known rust-gpu vertex and fragment entry points.

Functions

reload_materials

RustGpuBuilderOutput asset event handler.

Type Definitions

EntryPointConstants

A set of entry point constants for use with the EntryPoint trait.

EntryPointName

An entry point name for use with the EntryPoint trait.

EntryPointParameters

A set of entry point compile parameters for use with the EntryPoint trait.

EntryPointTypes

A set of entry point constants for use with the EntryPoint trait.