Module bevy_rust_gpu::prelude

Re-exports

pub use crate::*;

pub use crate::entry_point_export::*;

Structs

BuilderOutputPlugin

RustGpu

Extends a Material with rust-gpu shader support.

RustGpuArtifact

Asset containing loaded rust-gpu shaders and entry point metadata.

RustGpuArtifacts

RustGpuKey

RustGpu pipeline key.

RustGpuMaterial2dPlugin

Configures backend [Material2d] support for RustGpu<M>.

RustGpuMaterialPlugin

Configures backend [Material] support for RustGpu<M>.

RustGpuPlugin

Main Rust-GPU plugin.

RustGpuSettings

Type-level RustGpu material settings

Enums

RustGpuModules

Module shader handle container.

Statics

RUST_GPU_ARTIFACTS

Static container for RustGpuArtifacts to allow access from Material::specialize

Traits

EntryPoint

A rust-gpu entry point for use with RustGpuMaterial.

RustGpuMaterial

A [Material] type with statically-known rust-gpu vertex and fragment entry points.

Functions

builder_output_events

RustGpuBuilderOutput asset event handler.

reload_materials

RustGpuBuilderOutput asset event handler.

Type Definitions

EntryPointConstants

A set of entry point constants for use with the EntryPoint trait.

EntryPointName

An entry point name for use with the EntryPoint trait.

EntryPointParameters

A set of entry point compile parameters for use with the EntryPoint trait.

EntryPointTypes

A set of entry point constants for use with the EntryPoint trait.