Trait bevy_rust_gpu::EntryPoint

pub trait EntryPoint: 'static + Send + Sync {
    const NAME: &'static str;
    const PARAMETERS: EntryPointParameters;

    fn permutation(shader_defs: &Vec<String>) -> Vec<String>  { ... }
    fn build(shader_defs: &Vec<String>) -> String { ... }
}

A rust-gpu entry point for use with RustGpuMaterial.

Required Associated Constants

const NAME: &'static str

The entry point’s base function name, including module path

const NAME: EntryPointName = "mesh::entry_points::vertex";

const PARAMETERS: EntryPointParameters

Mapping from bevy shader defs to permutate-macro parameters.

const PARAMETERS: EntryPointParameters = &[
    (&[("VERTEX_TANGENTS", "some")], "none"),
    (&[("VERTEX_COLORS", "some")], "none"),
    (&[("SKINNED", "some")], "none"),
];

Provided Methods

fn permutation(shader_defs: &Vec<String>) -> Vec<String>

Constructs a permutation set from the provided shader defs

fn build(shader_defs: &Vec<String>) -> String

Build an entry point name from the provided shader defs