Crate bevy_rust_gpu

bevy-rust-gpu

A set of bevy plugins supporting the use of rust-gpu shaders.

Features include hot-reloading, metadata-based entry point validation, and active entry point export.

Can be used in conjunction with rust-gpu-builder and permutate-macro to drive a real-time shader recompilation pipeline.

Modules

bevy_pbr_rust

bevy-pbr-rust-backed RustGpuMaterial implementation for StandardMaterial.

entry_point_export

Adds support for exporting the active entry point set to a .json file.

prelude

shader_meta

Adds support for loading .spv.json metadata.

systems

Plugin systems.

Structs

ChangedShaders

A resource to track rust-gpu shaders that have been reloaded on a given frame

RustGpu

Extends a Material with rust-gpu shader support.

RustGpuKey

RustGpu pipeline key.

RustGpuMaterialPlugin

Configures backend support for RustGpu<M>.

RustGpuPlugin

Enforces WGPU limitations required by rust-gpu, and runs initial backend setup.

RustGpuSettings

Type-level RustGpu material settings

Traits

EntryPoint

A rust-gpu entry point for use with RustGpuMaterial.

LoadRustGpuShader

Loads a Shader asset, and optionally its metadata if the corresponding feature flag is enabled.

RustGpuMaterial

A [Material] type with statically-known rust-gpu vertex and fragment entry points.

Type Definitions

EntryPointConstants

A set of entry point constants for use with the EntryPoint trait.

EntryPointName

An entry point name for use with the EntryPoint trait.

EntryPointParameters

A set of entry point compile parameters for use with the EntryPoint trait.