Trait bevy_rust_gpu::EntryPoint

pub trait EntryPoint: 'static + Send + Sync {
    const NAME: &'static str;

    fn parameters() -> EntryPointParameters { ... }
    fn constants() -> EntryPointConstants { ... }
    fn types() -> EntryPointTypes { ... }
    fn permutation(shader_defs: &Vec<ShaderDefVal>) -> Vec<String>  { ... }
    fn filter_constants(shader_defs: &Vec<ShaderDefVal>) -> Vec<ShaderDefVal>  { ... }
    fn build(shader_defs: &Vec<ShaderDefVal>) -> String { ... }
}

A rust-gpu entry point for use with RustGpuMaterial.

Required Associated Constants

const NAME: &'static str

The entry point’s base function name, including module path

const NAME: EntryPointName = "mesh::entry_points::vertex";

Provided Methods

fn parameters() -> EntryPointParameters

fn constants() -> EntryPointConstants

fn types() -> EntryPointTypes

fn permutation(shader_defs: &Vec<ShaderDefVal>) -> Vec<String>

Constructs a permutation set from the provided shader defs

fn filter_constants(shader_defs: &Vec<ShaderDefVal>) -> Vec<ShaderDefVal>

fn build(shader_defs: &Vec<ShaderDefVal>) -> String

Build an entry point name from the provided shader defs